+++ .gitmessage
@@ -1,4 +1,4 @@
-Feat: 프론트엔드 스켈레톤 작동 가능하게 완성
+Refactor: 관심사 분리 패턴 적용 - models와 schemas 분리
 # #로 시작하는 줄은 주석입니다. 주석은 지우지 마세요!!
 # <Type>(<scope>): <제목> (가능하면 50자 이내, 마침표 없이)
 # -----type 목록-----
